services/mqtt-listener/requirements.txt
paho-mqtt==1.6.1
mysql-connector-python==8.3.0
